# golang-example/swagger/model_todo.go
/*
 * TO-DO list
 *
 * This is an example server for TO-DO list.  This is part of the collection of articles dedicated to the development of RESTful API. To know more visit the [repository](https://github.com/dandpz/restfulapi-howto) containing the code for these examples and more.
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Todo struct {

	Id int64 `json:"id,omitempty"`

	Name string `json:"name,omitempty"`

	Description string `json:"description,omitempty"`

	DueDate string `json:"dueDate,omitempty"`
	// to-do Status
	Status string `json:"status,omitempty"`
}